requirements.txt
blinker==1.9.0
boto3==1.37.7
botocore==1.37.7
certifi==2025.1.31
click==8.1.8
dnspython==2.7.0
et_xmlfile==2.0.0
Flask==3.1.0
Flask-Mail==0.10.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
MarkupSafe==3.0.2
openpyxl==3.1.5
packaging==24.2
pymongo==4.11.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.11.4
six==1.17.0
urllib3==2.3.0
Werkzeug==3.1.3
